Tweaked: Better chunk utlization Fixed: Defragmentation on chunkless archetypes

+++ include/gaia/ecs/chunk_allocator.h
@@ -219,7 +219,7 @@ namespace gaia {
 				ChunkAllocatorImpl& operator=(ChunkAllocatorImpl&&) = delete;
 				ChunkAllocatorImpl& operator=(const ChunkAllocatorImpl&) = delete;
 
-				static uint32_t GetMemoryBlockSize(uint32_t sizeType) {
+				static uint16_t GetMemoryBlockSize(uint32_t sizeType) {
 					return sizeType != 0 ? MaxMemoryBlockSize : MaxMemoryBlockSize / 2;
 				}
